Grader PC

+++ 29_oddeven.c
@@ -1,7 +1,7 @@
 #include<stdio.h>
 int main(){
-  double num;
-  scanf("%lf", &num);
+  int num;
+  scanf("%d", &num);
   if(num%2==0){
     printf("even");
   }
